refactor: trasnfere pageContainer to service
